+++ crates/editor/src/actions/indent.rs
@@ -4,7 +4,7 @@ use crate::{common::indent::determine_indent, editor::Editor, server::ClientId};
 
 #[action("Detect indentation")]
 fn detect_indent(editor: &mut Editor, id: ClientId) {
-    const MAX: usize = 1024 * 64; // 64kb
+    const MAX: u64 = 1024 * 64; // 64kb
 
     let (win, buf) = editor.win_buf_mut(id);
     let len = buf.len();
